externalize utils functions in webant.api
